+++ src/Viewer.tsx
@@ -39,9 +39,16 @@ interface RenderViewerProps {
     zoom(level: number | SpecialZoomLevel): void;
 }
 
+export interface DocumentLoadEvent {
+    doc: PdfJs.PdfDocument;
+}
 export interface PageChangeEvent {
     currentPage: number;
     doc: PdfJs.PdfDocument;
+}
+export interface ZoomEvent {
+    doc: PdfJs.PdfDocument;
+    scale: number;
 }
 
 export type RenderViewer = (props: RenderViewerProps) => React.ReactElement;
@@ -64,9 +71,9 @@ interface ViewerProps {
     renderPage?: RenderPage;
     // The text selection mode
     selectionMode?: SelectionMode;
-    onDocumentLoad?(doc: PdfJs.PdfDocument): void;
+    onDocumentLoad?(e: DocumentLoadEvent): void;
     onPageChange?(e: PageChangeEvent): void;
-    onZoom?(doc: PdfJs.PdfDocument, scale: number): void;
+    onZoom?(e: ZoomEvent): void;
 }
 
 const Viewer: React.FC<ViewerProps> = ({