+++ CHANGELOG.md
@@ -19,6 +19,15 @@ const handlePageChange = (e: PageChangeEvent) => {
 />
 ~~~
 
+**Breaking changes**
+
+The parameters of `onDocumentLoad` and `onZoom` are changed as following:
+
+| v1.6.0                    | v1.7.0                                |
+|---------------------------|---------------------------------------|
+| `onDocumentLoad(doc)`     | `onDocumentLoad({ doc } )`            |
+| `onZoom(doc, scale)`      | `onDocumentLoad({ doc, scale } )`     |
+
 ## v1.6.0
 
 **New features**